Fixed issues with negative signs throughout the controls model.  This required more than just removing negative signs, but after many changes the code is in good shape.  Changes were mode in the controls and plant model.
The coordinate system is +x for driving away from the driver station and +y for driving to the left of the driver station.  Positive rotation is counter clockwise as viewed from the top.

Various model documentation cleanup in many of the subsystems.

Updated odometry and dropped the calculations into the top level of the plant model to verify the calculations were correct.  Needed to make a small change to the logic, but it is now working very well.  Next step is to integrate the odometry code into the controls model in the appropriate location.