Getting the swerve puppy robot up and running again so it is performing as well as it did in July 2023.
Janelyn Anderson and Randy Anderson

Correctly mapped joysticks to the field coordinate system.
Field_x = -Joystick_y
Field_y = -Joystick_x

Note: Field_x is away from the driver station and Field_y is to the left of the driver station.

Wheel calibration improvement:
In the 'test' game state c++ code added joystick button inputs to make it easier to calibrate the wheels.

Wheel calibration procedure:
1. Go into 'Test' game state
2. Push the Wheel Off button (left/drive joystick button #6).  This sets steering duty cycle to 0 and puts them into coast mode.
3. Alight the wheels with gears facing RIGHT.  If this is done backwards then robot will steer in the opposite directions.
4. Push the Calibrate button (left/drive joystick button #8).
5. (optional) Push the Wheel On button (left/drive joystick button #11).

Added GameInitValues():
 * Changes variables when a game state changes, this is a safety measure for when a game is played and the n a robot is disabled and re-enabled, this makes sure there are no states carried over between 'enables'

Debugged and resolved an issue with one or two drive motors being commanded slower than others and thus resulting in driving on an arc instead of straight.  Root cause of the issue was with the drive motor speed reduction logic added based on simulation work, but which had not been tested on the robot.   The drive motor reduction logic slows down the drive motor when there is a large steering module error until the error gets back close to 0 in order to corner with less overshoot.  The angle error signal was getting out of the expected range and forcing a 0.05 factor drive speed reduction factor. To resolve this issue a modulus function was needed on the angle error signal and limited to +/- 90 degrees (in radians).

Stopped the robot from very slowly rotating while in absolute steering mode.  The issue wasn't due to gyro drift, but by a very small absolute steering PID value hanging at a non-zero value.  Set the integral gain to 0 and changed the dead zone logic.  Remove the Simulink block and replaced with a simple magnitude check and switch to force a 0 steering drive command when the PID output is close to 0.

Cleaned and refactored other parts of the swerve drive code in Simulink including renaming library blocks to be more descriptive, deleting some unused library blocks, and adding annotations.  Some of these changes were done AFTER running the robot, so this commit is not exactly what was run at the shop.

Added a smart dashboard layout to source control for debugging swerve drive..

Small changes in hand-code; got rid of all default values in SwerveDrive.cpp and moved them to constants:
 * drive mode reversals
 * steer mode reversals
 * default states

Moved debug buttons to Test Mode