William Anderson and Randy Anderson
Integrated Photon Vision Pose inputs into Simulink and added the Kalman filter which is working in the 3D simulation.  Unfortunately, we cannot get VS Code to build successfully due to a missing hand code library (emmintrin.h) that doesn't seem to be present anywhere in our system.  For now, commented out the Kalman filter in the Code_Gen_Model and rebuilt everything.  Might try to build our own simple Kalman filter instead of using the built-in one from Matlab as it seems to be using s-functions that are causing our VS Code to not build.