Randy Anderson
Replaced Extended Kalman Filter with Kalman Filter.  The EKF is only needed for nonlinear systems (which ours is not based on how we formulated the inputs and measurements) and multi-rate inputs.  The multi-rate might be nice for the camera measurement in the future, but at this time all of Simulink is on one periodic task so it is single rate.  We should be fine with the standard KF.

Resolved the code generation error by changing the Simulink configuration setting for Hardware Implementation from Intel-x86-64 to ARM Compatible ARM-7.  This removed the need for the emmintrin.h file which the RoboRio ARM processor doesn't support.

Simulation results look good, code generates, and code compiles.  Ready to test.